requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
imbalanced-learn==0.13.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lightgbm==4.6.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
newrelic==10.14.0
nltk==3.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pillow==11.3.0
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==5.0.0
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.35.0
